split ethash algorithm into dag generation and mining part
