+++ libethash-cuda/dag_globals.h
@@ -1,8 +1,11 @@
 #pragma once
 
+#include "dag_generation_kernel.h"
+
 __constant__ uint32_t d_dag_size;
 __constant__ hash128_t* d_dag;
 __constant__ uint32_t d_light_size;
 __constant__ hash64_t* d_light;
 __constant__ hash32_t d_header;
 __constant__ uint64_t d_target;
+
