+++ libethash-cuda/fnv.cuh
