+++ libethash-cuda/progpow/ProgPowMiner.cpp
@@ -328,9 +328,7 @@ bool CUDAMiner::s_noeval = false;
 
 bool CUDAMiner::cuda_init(
 	size_t numDevices,
-	ethash_light_t _light,
-	uint8_t const* _lightData,
-	uint64_t _lightBytes,
+	int epoch,
 	unsigned _deviceId,
 	bool _cpyToHost,
 	uint8_t* &hostDAG,
@@ -355,9 +353,13 @@ bool CUDAMiner::cuda_init(
 		m_search_buf = new volatile search_results *[s_numStreams];
 		m_streams = new cudaStream_t[s_numStreams];
 
-		uint64_t dagBytes = ethash_get_datasize(_light->block_number);
-		uint32_t dagElms   = (unsigned)(dagBytes / ETHASH_MIX_BYTES);
-		uint32_t lightWords = (unsigned)(_lightBytes / sizeof(node));
+		const auto& context = ethash::managed::get_epoch_context(epoch);
+        	const auto lightNumItems = context.light_cache_num_items;
+        	const auto lightSize = ethash::get_light_cache_size(lightNumItems);
+        	const auto dagNumItems = context.full_dataset_num_items;
+        	const auto dagSize = ethash::get_full_dataset_size(dagNumItems);
+		uint32_t dagElms   = (unsigned)(dagNumItems / ETHASH_MIX_BYTES);
+		uint32_t lightWords = (unsigned)(lightSize / sizeof(node));
 
 		CUDA_SAFE_CALL(cudaSetDevice(m_device_num));
 		cudalog << "Set Device to current";
